File Practice

+++ file2.py
@@ -1,2 +1,5 @@
 import os
-os.remove("file2.txt")
+if os.path.exists("file2.txt"):
+    os.remove("file2.txt")
+else:
+    print("File doesn't exist!")
